requirements.txt
Flask
SQLAlchemy
psycopg2-binary==2.9.9
flask_sqlalchemy
Flask-Migrate
flask-cors
instaloader
python-dotenv
selenium
bs4
pandas 
chromedriver-py
stdiomask
tweepy 
streamlit
seaborn
matplotlib
config
plotly
tweepy
schedule
insta-scrape
